# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "JavInfoApi",
    "description": "媒体元数据查询 API，基于 r18.dev 数据库（PostgreSQL）",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发服务器"
    }
  ],
  "paths": {
...
    "/api/v1/videos/batch": {
      "post": {
        "summary": "批量获取视频详情",
        "description": "根据 content_id 列表批量获取视频完整信息（含关联数据），最多100个",
        "operationId": "batchGetVideos",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ids"],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": { "type": "string" },
                    "maxItems": 100,
                    "description": "content_id 列表"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/VideoDetailResponse" }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/videos/lookup": {
      "post": {
        "summary": "批量番号查找视频",
        "description": "根据 dvd_id 列表批量查找视频（自动处理横杠和大小写），最多100个。返回 dvd_id -> video 的映射",
        "operationId": "batchLookupVideos",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["dvd_ids"],
                "properties": {
                  "dvd_ids": {
                    "type": "array",
                    "items": { "type": "string" },
                    "maxItems": 100,
                    "description": "dvd_id 列表（支持 ABC-001 或 ABC001 格式）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功，返回 {dvd_id: video} 映射",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { "$ref": "#/components/schemas/VideoDetailResponse" }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VideoDetailResponse": {
        "type": "object",
        "properties": {
          "content_id": { "type": "string" },
          "dvd_id": { "type": "string" },
          "title_en": { "type": "string" },
          "title_ja": { "type": "string" },
          "comment_en": { "type": "string" },
          "comment_ja": { "type": "string" },
          "runtime_mins": { "type": "integer" },
          "release_date": { "type": "string", "format": "date" },
          "sample_url": { "type": "string" },
          "maker_id": { "type": "integer" },
          "label_id": { "type": "integer" },
          "series_id": { "type": "integer" },
          "jacket_full_url": { "type": "string" },
          "jacket_thumb_url": { "type": "string" },
          "gallery_thumb_first": { "type": "string" },
          "gallery_thumb_last": { "type": "string" },
          "site_id": { "type": "integer" },
          "service_code": { "type": "string" },
          "maker": { "$ref": "#/components/schemas/Maker" },
          "label": { "$ref": "#/components/schemas/Label" },
          "series": { "$ref": "#/components/schemas/Series" },
          "actresses": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ActressWithOrdinality" }
          },
          "categories": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Category" }
          }
        }
      },
...
      "ActressWithOrdinality": {
        "type": "object",
        "description": "视频详情中的演员信息，包含排序序号",
        "properties": {
          "id": { "type": "integer" },
          "name_romaji": { "type": "string" },
          "name_kanji": { "type": "string" },
          "name_kana": { "type": "string" },
          "image_url": { "type": "string" },
          "ordinality": { "type": "integer", "description": "演员在该视频中的排序序号" }
        }
      },
      "Maker": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_en": { "type": "string" },
          "name_ja": { "type": "string" }
        }
      },
...
      "Label": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_en": { "type": "string" },
          "name_ja": { "type": "string" }
        }
      },
...
      "Series": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_en": { "type": "string" },
          "name_ja": { "type": "string" }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name_en": { "type": "string" },
          "name_ja": { "type": "string" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  }
}